+++ xamarin/Framework/LiferayScreens.Android/Additions/Auth/Login/LoginScreenlet.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Com.Liferay.Mobile.Screens.Auth.Login
+{
+    public partial class LoginScreenlet
+    {
+        protected override Java.Lang.Object CreateInteractor(string p0)
+        {
+            throw new NotImplementedException();
+        }
+
+        protected override void OnUserAction(string p0, Java.Lang.Object p1, params Java.Lang.Object[] p2)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}