+++ xamarin/Framework/LiferayScreens.Android/Additions/Ddl/Form/Model/StringWithOptionsField.cs
@@ -1,0 +1,22 @@
+﻿using System;
+
+namespace Com.Liferay.Mobile.Screens.Ddl.Model
+{
+    public partial class StringWithOptionsField
+    {
+        protected override string ConvertToFormattedString(Java.Lang.Object p0)
+        {
+            throw new NotImplementedException();
+        }
+
+        protected override Java.Lang.Object ConvertFromString(string p0)
+        {
+            throw new NotImplementedException();
+        }
+
+        protected override string ConvertToData(Java.Lang.Object p0)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}