+++ xamarin/Framework/LiferayScreens.Android/Additions/Viewset/Default/DDL/Form/CustomRatingNumberView.cs
@@ -1,0 +1,20 @@
+﻿using Com.Liferay.Mobile.Screens.Ddl.Model;
+
+namespace Com.Liferay.Mobile.Screens.Viewsets.Defaultviews.Ddl.Form.Fields
+{
+    public partial class CustomRatingNumberView
+    {
+        public global::Java.Lang.Object Field
+        {
+            get
+            {
+                return this.FieldProperty;
+            }
+
+            set
+            {
+                this.FieldProperty = (NumberField)value;
+            }
+        }
+    }
+}