+++ xamarin/Framework/LiferayScreens.Android/Additions/Viewset/Default/DDL/Form/DDLFieldRadioView.cs
@@ -1,0 +1,14 @@
+﻿namespace Com.Liferay.Mobile.Screens.Viewsets.Defaultviews.Ddl.Form.Fields
+{
+    public partial class DDLFieldRadioView
+    {
+        public global::Java.Lang.Object Field
+        {
+            get
+            {
+                return this.FieldProperty;
+            }
+            set {}
+        }
+    }
+}